Activity hub improvements - hide last week and add current month stats instead if tag has a monthly frequency target reflecting stats based on time frequency, allow decimal values for tag target value, only concatenate event activity with hub activity if event activity tag is targeted by hours fix. When adding activity to hub make input element instead of select if target unit is not by hours and allow adding double values.

+++ Appology/Model/Calendar/Event.cs
@@ -36,7 +36,7 @@ namespace Appology.MiCalendar.Model
         [DbIgnore]
         public TimeFrequency? TargetFrequency { get; set; }
         [DbIgnore]
-        public int? TargetValue { get; set; }
+        public double? TargetValue { get; set; }
         [DbIgnore]
         public string TargetUnit { get; set; }
         [DbIgnore]
